Simplified CPP example
